Fixed the animation Issue with the C# script port.
Need to fix screen resolutions. There are still issues with text
placement. Should have first array be randomized. Might be a good way to
add to level progression. As for the animation fix, had to start a
coroutine as it was: yield return new WaitForSeconds(waitTime). Seems to
work just fine now. This completes the original port from UnityScript
and anything else added to this will be additional functionality.